+++ src/main/java/me/kevinboone/caztor/converters/Converter.java
@@ -11,6 +11,7 @@ package me.kevinboone.caztor.converters;
 import java.net.*;
 import java.io.*;
 import java.util.regex.Pattern;
+import java.nio.charset.Charset;
 import me.kevinboone.caztor.base.*;
 
 /** This interface is implemented by all something-to-HTML converters 
@@ -19,6 +20,6 @@ public interface Converter
   {
   /** Given a String s of whatever the content is, return a new String
       of HTML. */
-  public String toHtml (String s);
+  public String toHtml (byte[] s, Charset charset);
   }