+++ src/main/java/me/kevinboone/caztor/protocol/ZipfileURLStreamHandler.java
@@ -1,0 +1,35 @@
+/*============================================================================
+
+  Caztor
+
+  ZipfileURLStreamHandler
+
+  Copyright (c)2021-2026 Kevin Boone, GPLv3.0
+
+============================================================================*/
+package me.kevinboone.caztor.protocol; 
+
+import java.io.IOException;
+import java.net.URL;
+import java.net.URLConnection;
+import java.net.URLStreamHandler;
+
+/** A URLStreamHandler that instantiates ZipfileConnection. 
+*/
+public class ZipfileURLStreamHandler extends URLStreamHandler
+  {
+  @Override
+  protected URLConnection openConnection (URL url)
+  throws IOException
+    {
+    return new ZipfileConnection (url);
+    }
+
+  @Override
+  protected int getDefaultPort ()
+    {
+    return 0;
+    }
+  }
+
+